cleanup stripe checkout
