+++ frontend/src/components/StripeCheckoutForm/StripeCheckoutForm.tsx
@@ -1,20 +1,29 @@
 import React, { FormEvent, useEffect, useState } from "react";
 
 import {
+  Elements,
   PaymentElement,
   useElements,
   useStripe,
 } from "@stripe/react-stripe-js";
-import { StripePaymentElementOptions } from "@stripe/stripe-js";
+import { StripePaymentElementOptions, loadStripe } from "@stripe/stripe-js";
+
+const prod = import.meta.env.VITE_PROD == "1";
+
+const stripePromise = loadStripe(
+  prod
+    ? "pk_live_51MePU5JR76QyQ6AvL89IwWk9yYLIH3ERANOGGCWgJOYeNCLe3gZkc610rqRQazf2anIVB5wz3ob05zleH0q4I1Jy00bztowdqI"
+    : "pk_test_51MePU5JR76QyQ6Avy732ksgL17hzuDKNXVtI9zpBJrwzfcsshlp9QnygfuaVTKRptbjxts9V6GB1AQEmkacsYdq400LvhshcA4"
+);
 
 interface Props {
-  successURL: string;
   paymentIntentClientSecret: string;
+  successURL: string;
 }
 
 export function StripeCheckoutForm({
-  successURL,
   paymentIntentClientSecret,
+  successURL,
 }: Props) {
   const stripe = useStripe();
   const elements = useElements();
@@ -26,10 +35,6 @@ export function StripeCheckoutForm({
     if (!stripe) {
       return;
     }
-
-    // const paymentIntentClientSecret = new URLSearchParams(window.location.search).get(
-    //   "payment_intent_client_secret"
-    // );
 
     if (!paymentIntentClientSecret) {
       return;
@@ -86,20 +91,32 @@ export function StripeCheckoutForm({
       setMessage("An unexpected error occurred.");
     }
 
+    // TODO: error notification
     setIsLoading(false);
   };
 
   const paymentElementOptions: StripePaymentElementOptions = {
     layout: "tabs",
     paymentMethodOrder: ["card"],
-  }
+  };
+
+  const appearance = {
+    theme: "stripe",
+  } as const;
+
+  const options = {
+    clientSecret: paymentIntentClientSecret,
+    appearance,
+  } as const;
 
   return (
-    <form id="payment-form" onSubmit={handleSubmit}>
-      <PaymentElement id="payment-element" options={paymentElementOptions} />
+    <Elements options={options} stripe={stripePromise}>
+      <form id="payment-form" onSubmit={handleSubmit}>
+        <PaymentElement id="payment-element" options={paymentElementOptions} />
 
-      {/* Show any error or success messages */}
-      {message && <div id="payment-message">{message}</div>}
-    </form>
+        {/* Show any error or success messages */}
+        {message && <div id="payment-message">{message}</div>}
+      </form>
+    </Elements>
   );
 }
