+++ frontend/src/features/users/components/BuyCredits.tsx
@@ -4,29 +4,13 @@ import { Button } from "@/components/Button";
 import { FormDrawer } from "@/components/Form";
 
 import { PencilIcon } from "@heroicons/react/24/outline";
-import { loadStripe } from "@stripe/stripe-js";
 
 import { StripeCheckoutForm } from "@/components/StripeCheckoutForm";
 import { creditsPaymentIntent } from "@/features/projects/requests";
-import { Elements } from "@stripe/react-stripe-js";
 
 import { TextInput } from "@mantine/core";
+import { successURL } from "./StripePayment";
 
-// Make sure to call loadStripe outside of a component’s render to avoid
-// recreating the Stripe object on every render.
-// This is your test publishable API key.
-const stripePromise = loadStripe(
-  "pk_live_51MePU5JR76QyQ6AvL89IwWk9yYLIH3ERANOGGCWgJOYeNCLe3gZkc610rqRQazf2anIVB5wz3ob05zleH0q4I1Jy00bztowdqI"
-);
-
-const successURL = (credits: number) =>
-  import.meta.env.VITE_PROD == "1"
-    ? import.meta.env.VITE_PROD_URL +
-      `/app/profile?success=true&credits=${credits}`
-    : import.meta.env.VITE_DEV_URL +
-      `/app/profile?success=true&credits=${credits}`;
-
-// custom trigger button
 interface Props {
   btn?: React.ReactElement;
   title: string;
@@ -36,14 +20,6 @@ export const BuyCredits = ({ btn, title }: Props) => {
   const [clientSecret, setClientSecret] = useState("");
   const [credits, setCredits] = useState<string | number>("");
   const [price, setPrice] = useState<string | number>("");
-
-  const appearance = {
-    theme: "stripe",
-  } as const;
-  const options = {
-    clientSecret,
-    appearance,
-  } as const;
 
   return (
     <FormDrawer
@@ -71,11 +47,11 @@ export const BuyCredits = ({ btn, title }: Props) => {
       }
     >
       {!clientSecret && (
-        <div className="h-full flex flex-col items-center justify-center">
+        <div className="flex h-full flex-col items-center justify-center">
           <span className="text-2xl font-bold">Buy Credits</span>
           <form
             id="buy-credits"
-            className="w-64 mt-4"
+            className="mt-4 w-64"
             onSubmit={async (e) => {
               e.preventDefault();
               const creditsPaymentIntentSecret = await creditsPaymentIntent(
@@ -118,21 +94,11 @@ export const BuyCredits = ({ btn, title }: Props) => {
         </div>
       )}
       {clientSecret && (
-        <Elements options={options} stripe={stripePromise}>
-          <StripeCheckoutForm
-            successURL={successURL(Number(credits))}
-            paymentIntentClientSecret={clientSecret}
-          />
-        </Elements>
+        <StripeCheckoutForm
+          paymentIntentClientSecret={clientSecret}
+          successURL={successURL(Number(credits))}
+        />
       )}
-
-      {/* Show cost of credits, pretty */}
-
-      {/* <div className="mt-2">
-        <p className="">
-          {credits} credits = ${credits * 0.01}
-        </p>
-      </div> */}
     </FormDrawer>
   );
 };
